Switch v1/v2 paths
