+++ src/encoded/static/components/index.js
@@ -26,8 +26,8 @@ require('./platform');
 require('./search');
 require('./report');
 require('./matrix');
+require('./auditmatrixv1');
 require('./auditmatrix');
-require('./auditmatrixv2');
 require('./target');
 require('./publication');
 require('./pipeline');
